added test for current customer, added active attribute, added method for returning current customer
